% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{.cache_r4ds_slack_channels}
\alias{.cache_r4ds_slack_channels}
\title{Cache R4DS Slack channels}
\usage{
.cache_r4ds_slack_channels()
}
\value{
A dataframe with information about R4DS Slack channels, invisibly.
}
\description{
Set R4DS slack channels in the package \code{the} environment.
}
\keyword{internal}
