% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{.cache_r4ds_youtube_playlists}
\alias{.cache_r4ds_youtube_playlists}
\title{Cache R4DS YouTube playlists}
\usage{
.cache_r4ds_youtube_playlists(n)
}
\arguments{
\item{n}{How many playlists do we need? This should ideally be equal to the
number of active clubs.}
}
\value{
A character vector of playlist IDs, with titles as names, invisibly.
}
\description{
Set R4DS YouTube playlists in the package \code{the} environment.
}
\keyword{internal}
