% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{.fetch_r4ds_slack_channels}
\alias{.fetch_r4ds_slack_channels}
\title{Fetch R4DS Slack channels}
\usage{
.fetch_r4ds_slack_channels()
}
\value{
A dataframe with information about R4DS Slack channels.
}
\description{
Fetch R4DS Slack channels
}
\keyword{internal}
