% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{dslc_slack_channels}
\alias{dslc_slack_channels}
\title{Cache or Fetch DSLC Slack channels}
\usage{
dslc_slack_channels(refresh = FALSE)
}
\arguments{
\item{refresh}{Get fresh data?}
}
\value{
A dataframe with information about DSLC Slack channels.
}
\description{
Fetch public and private DSLC Slack channel information.
}
