% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{r4ds_slack_channels}
\alias{r4ds_slack_channels}
\title{Cache or Fetch R4DS Slack channels}
\usage{
r4ds_slack_channels(refresh = FALSE)
}
\arguments{
\item{refresh}{Get fresh data?}
}
\value{
A dataframe with information about R4DS Slack channels.
}
\description{
Fetch public and private R4DS Slack channel information.
}
