% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{r4ds_youtube_playlists}
\alias{r4ds_youtube_playlists}
\title{Cache or Fetch R4DS Slack channels}
\usage{
r4ds_youtube_playlists(n = 50L, refresh = FALSE)
}
\arguments{
\item{n}{How many playlists do we need? This should ideally be equal to the
number of active clubs.}

\item{refresh}{Get fresh data?}
}
\value{
A character vector of playlist IDs, with titles as names.
}
\description{
Fetch public and private R4DS Slack channel information.
}
