% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{remove_all_club_reminders}
\alias{remove_all_club_reminders}
\title{Remove Book Club Zoom reminders}
\usage{
remove_all_club_reminders(
  min_age_minutes = 55,
  token = slack_default_token(),
  slack_channels = dslc_slack_channels(token = token)
)
}
\arguments{
\item{min_age_minutes}{How old messages need to be (in minutes) to be
cleared.}

\item{token}{character, api token issued by slack}

\item{slack_channels}{A data.frame of Slack channels.}
}
\value{
NULL (invisibly)
}
\description{
Clear Zoom reminders from Book Club Slack channels.
}
