% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{remove_slack_reminders}
\alias{remove_slack_reminders}
\title{Remove Zoom reminders}
\usage{
remove_slack_reminders(
  channel_name,
  min_age_minutes = 55,
  max_msgs_to_check = Inf,
  token = slack_default_token(),
  slack_channels = dslc_slack_channels(token = token)
)
}
\arguments{
\item{channel_name}{The name of the channel from which reminders should be
removed.}

\item{min_age_minutes}{How old messages need to be (in minutes) to be
cleared.}

\item{max_msgs_to_check}{How many messages to fetch to check for Zoom
reminders.}

\item{token}{character, api token issued by slack}

\item{slack_channels}{A data.frame of Slack channels.}
}
\value{
NULL (invisibly)
}
\description{
Remove Zoom reminders from Slack channels.
}
\keyword{internal}
