% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the.R
\name{reset_the}
\alias{reset_the}
\title{Reset the the environment}
\usage{
reset_the()
}
\value{
\code{NULL} (invisibly).
}
\description{
\code{the} is an environment inside this package. This function resets it.
}
