% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookclubcron-package.R
\docType{package}
\name{bookclubcron-package}
\alias{bookclubcron}
\alias{bookclubcron-package}
\title{bookclubcron: Automated Tasks for R4DS Book Clubs}
\description{
Several tasks need to run multiple times per day to keep book clubs running.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/bookclubcron}
  \item \url{https://r4ds.github.io/bookclubcron/}
  \item Report bugs at \url{https://github.com/r4ds/bookclubcron/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
