% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{.cache_dslc_slack_channels}
\alias{.cache_dslc_slack_channels}
\title{Cache DSLC Slack channels}
\usage{
.cache_dslc_slack_channels()
}
\value{
A dataframe with information about DSLC Slack channels, invisibly.
}
\description{
Set DSLC slack channels in the package \code{the} environment.
}
\keyword{internal}
