% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{.check_can_upload}
\alias{.check_can_upload}
\title{Check YouTube quota space}
\usage{
.check_can_upload()
}
\value{
If there's room, \code{TRUE} invisibly (error otherwise).
}
\description{
Check the "dev-monitoring" R4DS Slack channel for warnings about quota usage.
}
\keyword{internal}
