% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{.fetch_dslc_slack_channels}
\alias{.fetch_dslc_slack_channels}
\title{Fetch DSLC Slack channels}
\usage{
.fetch_dslc_slack_channels()
}
\value{
A dataframe with information about DSLC Slack channels.
}
\description{
Fetch DSLC Slack channels
}
\keyword{internal}
