% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{.fetch_dslc_youtube_playlists}
\alias{.fetch_dslc_youtube_playlists}
\title{Fetch DSLC YouTube Playlists}
\usage{
.fetch_dslc_youtube_playlists(n)
}
\arguments{
\item{n}{How many playlists do we need? This should ideally be equal to the
number of active clubs.}
}
\value{
A character vector of playlist IDs, with titles as names.
}
\description{
Fetch DSLC YouTube Playlists
}
\keyword{internal}
