% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{dslc_slack_channels}
\alias{dslc_slack_channels}
\title{Cache or Fetch DSLC Slack channels}
\usage{
dslc_slack_channels(refresh = FALSE, token = slack_default_token())
}
\arguments{
\item{refresh}{Get fresh data?}

\item{token}{character, api token issued by slack}
}
\value{
A dataframe with information about DSLC Slack channels.
}
\description{
Fetch public and private DSLC Slack channel information.
}
