% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{dslc_youtube_playlists}
\alias{dslc_youtube_playlists}
\title{Cache or Fetch DSLC YouTube Playlists}
\usage{
dslc_youtube_playlists(n = 50L, refresh = FALSE)
}
\arguments{
\item{n}{How many playlists do we need? This should ideally be equal to the
number of active clubs.}

\item{refresh}{Get fresh data?}
}
\value{
A character vector of playlist IDs, with titles as names.
}
\description{
Fetch DSLC YouTube playlist information.
}
